Introduced track accessors and use them all over the code. Added templated extraction of SIMD vectors from track AOS for some track data members.

+++ vecprot_v2/inc/Basket.h
@@ -198,7 +198,7 @@ public:
    */
   VECCORE_ATT_HOST_DEVICE
   GEANT_FORCE_INLINE
-  int size() { return fTracks.size(); }
+  size_t size() { return fTracks.size(); }
 
   /**
    * @brief Print the basket content
